requirements.txt
streamlit
opencv-python
numpy
Pillow
requests